Move TlcConfigParser into infra.passes
This is really a configuration step, and we need access to these data
structures during the configuration/option-setting phase.
