+++ mod-infra/src/main/scala/at/forsyte/apalache/infra/tlc/config/TlcConfigTokenReader.scala
@@ -1,4 +1,4 @@
-package at.forsyte.apalache.io.tlc.config
+package at.forsyte.apalache.infra.tlc.config
 
 import scala.util.parsing.input.{NoPosition, Position, Reader}
